+++ Data/Debt.cs
@@ -11,8 +11,8 @@ namespace marpha.Data
         public int DebtId { get; set; }
         public string DebtName { get; set; }
         public string DebtDescription { get; set; }
-        public double PrincipleAmount { get; set; }
-        public double DebtAmount { get; set; }
+        public decimal PrincipleAmount { get; set; }
+        public decimal DebtAmount { get; set; }
         public DateTime DebtDate { get; set; }
         public DateTime DebtDueDate { get; set; }
         public string DebtStatus { get; set; }
